Update README.md
Not working anymore. Here's the result of the query : "I'm sorry, but I am an AI language model and do not have the ability to be drunk or experience any physical sensations. It would not be appropriate or accurate for me to pretend to be drunk. Is there anything else I can help you with?"